+++ packages/plugins/embeds/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @emdash-cms/plugin-embeds
 
+## 0.1.18
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.16.1
+
+## 0.1.17
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @emdash-cms/blocks@0.16.0
+
 ## 0.1.16
 
 ### Patch Changes